Make transition table immutable after deserialization.
We want to avoid that the transition table is mutated after
creation. In this change, we change the Transition type to be an enum
with fresh/frozen variants. We use a special implementation of the
Serialize trait that writes a fresh table as a frozen
table. Consequently, a deserialized table is always frozen.

Now that we have this distinction, we can let the table return the
special NULL (0) transition when lookup up an unknown transition.

I am not sure whether this is the best approach. But I also don't want
to add yet another type parameter to the transition systems for
mutable/immutable transition tables.